requirements.txt
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
requests==2.32.3
python-telegram-bot==13.15
python-dotenv>=0.21.0
urllib3==1.26.15
standard-imghdr==3.13.0
